requirements.txt
pelican~=4.8.0
pelican-webassets~=2.0.0
markdown~=3.4.1
pyyaml~=6.0
pystache
more_itertools~=9.0
bleach~=6.0
requests~=2.28.0
urllib3~=1.26.0
webassets~=2.0
pyscss~=1.4.0
cssmin~=0.2.0
livereload~=2.6.3
invoke~=2.0.0
git+https://github.com/webosbrew/pelican-theme-webosbrew.git@v1.0.4#egg=pelican-theme-webosbrew